requirements.txt
Flask>=2.0
google-cloud-storage>=2.0.0
gunicorn
pytest
